+++ packages/xontable/src/components/index.ts
@@ -2,3 +2,4 @@ export { ColumnFilterMenu } from "./ColumnFilterMenu";
 export { SelectMenu } from "./SelectMenu";
 export { XOnTableGrid } from "./XOnTableGrid";
 export { XOnTableHeader } from "./XOnTableHeader";
+export { XOnTableStatusBar } from "./XOnTableStatusBar";